+++ Amazon.QLDB.Driver.Tests/BufferedResultTests.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
  * the License. A copy of the License is located at
@@ -27,6 +27,8 @@ namespace Amazon.QLDB.Driver.Tests
         private static readonly IValueFactory valueFactory = new ValueFactory();
         private static Mock<IResult> mockResult;
         private static List<IIonValue> testList;
+        private static IOUsage testIO;
+        private static TimingInformation testTiming;
         private static BufferedResult result;
 
 
@@ -42,7 +44,11 @@ namespace Amazon.QLDB.Driver.Tests
                 valueFactory.NewInt(1),
                 valueFactory.NewInt(2)
             };
+            testIO = new IOUsage(1, 2);
+            testTiming = new TimingInformation(100);
             mockResult.Setup(x => x.GetAsyncEnumerator(It.IsAny<CancellationToken>())).Returns(new ValuesEnumerator(testList));
+            mockResult.Setup(x => x.GetConsumedIOs()).Returns(testIO);
+            mockResult.Setup(x => x.GetTimingInformation()).Returns(testTiming);
 
             result = await BufferedResult.BufferResult(mockResult.Object);
         }
@@ -52,6 +58,15 @@ namespace Amazon.QLDB.Driver.Tests
         {
             Assert.IsNotNull(result);
             mockResult.Verify(x => x.GetAsyncEnumerator(default), Times.Exactly(1));
+        }
+
+        [TestMethod]
+        public void TestBufferResultGetsMetrics()
+        {
+            Assert.AreEqual(testIO, result.GetConsumedIOs());
+            Assert.AreEqual(testTiming, result.GetTimingInformation());
+            mockResult.Verify(x => x.GetConsumedIOs(), Times.Exactly(1));
+            mockResult.Verify(x => x.GetTimingInformation(), Times.Exactly(1));
         }
 
         [TestMethod]