+++ Amazon.QLDB.Driver.Tests/QldbSessionTests.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
  * the License. A copy of the License is located at
@@ -150,7 +150,6 @@ namespace Amazon.QLDB.Driver.Tests
             {
                 await txn.Execute("testStatement");
                 throw new ArgumentException("some thing wrong");
-                return "result";
             };
 
             return new List<object[]>() {