+++ Amazon.QLDB.Driver.Tests/ResultTests.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
  * the License. A copy of the License is located at
@@ -27,19 +27,54 @@ namespace Amazon.QLDB.Driver.Tests
     public class ResultTests
     {
         private static Result result;
-        private static readonly Mock<Session> mockSession = new Mock<Session>(null, null, null, null, null);
-        private readonly MemoryStream memoryStream = new MemoryStream();
+        private static Mock<Session> mockSession;
+        private static readonly MemoryStream memoryStream = new MemoryStream();
+        private static readonly ValueHolder valueHolder = new ValueHolder
+        {
+            IonBinary = memoryStream,
+            IonText = "ionText"
+        };
+        private static readonly List<ValueHolder> valueHolderList = new List<ValueHolder> { valueHolder };
 
-        public IIonValue IonDatagram { get; private set; }
+        private static readonly long executeReads = 1;
+        private static readonly long executeWrites = 2;
+        private static readonly long executeTime = 100;
+        private static readonly IOUsage executeIO = new IOUsage
+        {
+            ReadIOs = executeReads,
+            WriteIOs = executeWrites
+        };
+        private static readonly TimingInformation executeTiming = new TimingInformation
+        {
+            ProcessingTimeMilliseconds = executeTime
+        };
+        private static readonly long fetchReads = 10;
+        private static readonly long fetchWrites = 20;
+        private static readonly long fetchTime = 1000;
+        private static readonly IOUsage fetchIO = new IOUsage
+        {
+            ReadIOs = fetchReads,
+            WriteIOs = fetchWrites
+        };
+        private static readonly TimingInformation fetchTiming = new TimingInformation
+        {
+            ProcessingTimeMilliseconds = fetchTime
+        };
 
         [TestInitialize]
         public void SetUp()
         {
-            var valueHolder = new ValueHolder { IonBinary = memoryStream, IonText = "ionText" };
-            var valueHolderList = new List<ValueHolder> { valueHolder };
-            var firstPage = new Page { NextPageToken = "hasNextPage", Values = valueHolderList };
+            var executeResult = new ExecuteStatementResult
+            {
+                FirstPage = new Page
+                {
+                    NextPageToken = "hasNextPage",
+                    Values = valueHolderList
+                }
+            };
 
-            result = new Result(mockSession.Object, "txnId", firstPage);
+            mockSession = new Mock<Session>(null, null, null, null, null);
+            result = new Result(mockSession.Object, "txnId", executeResult);
         }
 
         [TestMethod]
@@ -55,11 +90,14 @@ namespace Amazon.QLDB.Driver.Tests
         public async Task TestMoveNextWithOneNextPage()
         {
             var ms = new MemoryStream();
-            var valuHolderList = new List<ValueHolder> { new ValueHolder { IonBinary = ms, IonText = "ionText" } };
-            Page nextPage = new Page { NextPageToken = null, Values = valuHolderList };
+            var valueHolderList = new List<ValueHolder> { new ValueHolder { IonBinary = ms, IonText = "ionText" } };
+            var fetchPageResult = new FetchPageResult
+            {
+                Page = new Page { NextPageToken = null, Values = valueHolderList }
+            };
 
             mockSession.Setup(m => m.FetchPage(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
-                .ReturnsAsync(new FetchPageResult { Page = nextPage });
+                .ReturnsAsync(fetchPageResult);
 
             var results = result.GetAsyncEnumerator();
 
@@ -76,12 +114,18 @@ namespace Amazon.QLDB.Driver.Tests
         [TestMethod]
         public async Task TestMoveNextWithNoNextPage()
         {
-            Mock<Session> session = new Mock<Session>(null, null, null, null, null);
             var ms = new MemoryStream();
-            var valuHolderList = new List<ValueHolder> { new ValueHolder { IonBinary = ms, IonText = "ionText" } };
-            var firstPage = new Page { NextPageToken = null, Values = valuHolderList };
+            var valueHolderList = new List<ValueHolder> { new ValueHolder { IonBinary = ms, IonText = "ionText" } };
+            var executeResult = new ExecuteStatementResult
+            {
+                FirstPage = new Page
+                {
+                    NextPageToken = null,
+                    Values = valueHolderList
+                }
+            };
 
-            Result res = new Result(session.Object, "txnId", firstPage);
+            Result res = new Result(mockSession.Object, "txnId", executeResult);
             var results = res.GetAsyncEnumerator();
 
             int counter = 0;
@@ -91,7 +135,7 @@ namespace Amazon.QLDB.Driver.Tests
             }
 
             Assert.AreEqual(1, counter);
-            session.Verify(m => m.FetchPage(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Never);
+            mockSession.Verify(m => m.FetchPage(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()), Times.Never);
         }
 
         [TestMethod]
@@ -101,6 +145,170 @@ namespace Amazon.QLDB.Driver.Tests
 
             Assert.IsNotNull(results);
             Assert.IsTrue(await results.MoveNextAsync());
+        }
+
+        [TestMethod]
+        public async Task TestQueryStatsNullExecuteNullFetch()
+        {
+            var executeResult = GetExecuteResultNullStats();
+            var fetchPageResult = GetFetchResultNullStats();
+
+            mockSession.Setup(m => m.FetchPage(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(fetchPageResult);
+
+            var result = new Result(mockSession.Object, "txnId", executeResult);
+
+            var io = result.GetConsumedIOs();
+            var timing = result.GetTimingInformation();
+            Assert.IsNull(io.ReadIOs);
+            Assert.IsNull(io.WriteIOs);
+            Assert.IsNull(timing.ProcessingTimeMilliseconds);
+
+            var results = result.GetAsyncEnumerator();
+            while (await results.MoveNextAsync())
+            {
+                // Fetch the next page
+            }
+
+            io = result.GetConsumedIOs();
+            timing = result.GetTimingInformation();
+            Assert.IsNull(io.ReadIOs);
+            Assert.IsNull(io.WriteIOs);
+            Assert.IsNull(timing.ProcessingTimeMilliseconds);
+        }
+
+        [TestMethod]
+        public async Task TestQueryStatsNullExecuteHasFetch()
+        {
+            var executeResult = GetExecuteResultNullStats();
+            var fetchPageResult = GetFetchResultWithStats();
+
+            mockSession.Setup(m => m.FetchPage(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(fetchPageResult);
+
+            var result = new Result(mockSession.Object, "txnId", executeResult);
+
+            var io = result.GetConsumedIOs();
+            var timing = result.GetTimingInformation();
+            Assert.IsNull(io.ReadIOs);
+            Assert.IsNull(io.WriteIOs);
+            Assert.IsNull(timing.ProcessingTimeMilliseconds);
+
+            var results = result.GetAsyncEnumerator();
+            while (await results.MoveNextAsync())
+            {
+                // Fetch the next page
+            }
+
+            io = result.GetConsumedIOs();
+            timing = result.GetTimingInformation();
+            Assert.AreEqual(fetchReads, io.ReadIOs);
+            Assert.AreEqual(fetchWrites, io.WriteIOs);
+            Assert.AreEqual(fetchTime, timing.ProcessingTimeMilliseconds);
+        }
+
+        [TestMethod]
+        public async Task TestQueryStatsHasExecuteNullFetch()
+        {
+            var executeResult = GetExecuteResultWithStats();
+            var fetchPageResult = GetFetchResultNullStats();
+
+            mockSession.Setup(m => m.FetchPage(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(fetchPageResult);
+
+            var result = new Result(mockSession.Object, "txnId", executeResult);
+
+            var io = result.GetConsumedIOs();
+            var timing = result.GetTimingInformation();
+            Assert.AreEqual(executeReads, io.ReadIOs);
+            Assert.AreEqual(executeWrites, io.WriteIOs);
+            Assert.AreEqual(executeTime, timing.ProcessingTimeMilliseconds);
+
+            var results = result.GetAsyncEnumerator();
+            while (await results.MoveNextAsync())
+            {
+                // Fetch the next page
+            }
+
+            io = result.GetConsumedIOs();
+            timing = result.GetTimingInformation();
+            Assert.AreEqual(executeReads, io.ReadIOs);
+            Assert.AreEqual(executeWrites, io.WriteIOs);
+            Assert.AreEqual(executeTime, timing.ProcessingTimeMilliseconds);
+        }
+
+        [TestMethod]
+        public async Task TestQueryStatsHasExecuteHasFetch()
+        {
+            var executeResult = GetExecuteResultWithStats();
+            var fetchPageResult = GetFetchResultWithStats();
+
+            mockSession.Setup(m => m.FetchPage(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(fetchPageResult);
+
+            var result = new Result(mockSession.Object, "txnId", executeResult);
+
+            var io = result.GetConsumedIOs();
+            var timing = result.GetTimingInformation();
+            Assert.AreEqual(executeReads, io.ReadIOs);
+            Assert.AreEqual(executeWrites, io.WriteIOs);
+            Assert.AreEqual(executeTime, timing.ProcessingTimeMilliseconds);
+
+            var results = result.GetAsyncEnumerator();
+            while (await results.MoveNextAsync())
+            {
+                // Fetch the next page
+            }
+
+            io = result.GetConsumedIOs();
+            timing = result.GetTimingInformation();
+            Assert.AreEqual(executeReads + fetchReads, io.ReadIOs);
+            Assert.AreEqual(executeWrites + fetchWrites, io.WriteIOs);
+            Assert.AreEqual(executeTime + fetchTime, timing.ProcessingTimeMilliseconds);
+        }
+
+        private ExecuteStatementResult GetExecuteResultNullStats()
+        {
+            return new ExecuteStatementResult
+            {
+                FirstPage = new Page
+                {
+                    NextPageToken = "hasNextPage",
+                    Values = valueHolderList
+                }
+            };
+        }
+
+        private ExecuteStatementResult GetExecuteResultWithStats()
+        {
+            return new ExecuteStatementResult
+            {
+                FirstPage = new Page
+                {
+                    NextPageToken = "hasNextPage",
+                    Values = valueHolderList
+                },
+                ConsumedIOs = executeIO,
+                TimingInformation = executeTiming
+            };
+        }
+
+        private FetchPageResult GetFetchResultNullStats()
+        {
+            return new FetchPageResult
+            {
+                Page = new Page { NextPageToken = null, Values = valueHolderList }
+            };
+        }
+
+        private FetchPageResult GetFetchResultWithStats()
+        {
+            return new FetchPageResult
+            {
+                Page = new Page { NextPageToken = null, Values = valueHolderList },
+                ConsumedIOs = fetchIO,
+                TimingInformation = fetchTiming
+            };
         }
     }
 }