+++ Amazon.QLDB.Driver/result/BufferedResult.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
  * the License. A copy of the License is located at
@@ -26,15 +26,21 @@ namespace Amazon.QLDB.Driver
     public class BufferedResult : IResult
     {
         private readonly List<IIonValue> values;
+        private readonly IOUsage consumedIOs;
+        private readonly TimingInformation timingInformation;
 
         /// <summary>
         /// Prevents a default instance of the <see cref="BufferedResult"/> class from being created.
         /// </summary>
         ///
         /// <param name="values">Buffer values.</param>
-        private BufferedResult(List<IIonValue> values)
+        /// <param name="consumedIOs">IOUsage statistics.</param>
+        /// <param name="timingInformation">TimingInformation statistics.</param>
+        private BufferedResult(List<IIonValue> values, IOUsage consumedIOs, TimingInformation timingInformation)
         {
             this.values = values;
+            this.consumedIOs = consumedIOs;
+            this.timingInformation = timingInformation;
         }
 
         /// <summary>
@@ -52,7 +58,7 @@ namespace Amazon.QLDB.Driver
                 values.Add(value);
             }
 
-            return new BufferedResult(values);
+            return new BufferedResult(values, result.GetConsumedIOs(), result.GetTimingInformation());
         }
 
         /// <summary>
@@ -83,6 +89,26 @@ namespace Amazon.QLDB.Driver
             public ValueTask<bool> MoveNextAsync() => new ValueTask<bool>(this.valuesEnumerator.MoveNext());
 
             public ValueTask DisposeAsync() => default;
+        }
+
+        /// <summary>
+        /// Gets the current query statistics for the number of read IO requests. The statistics are stateful.
+        /// </summary>
+        ///
+        /// <returns>The current IOUsage statistics.</returns>
+        public IOUsage GetConsumedIOs()
+        {
+            return this.consumedIOs;
+        }
+
+        /// <summary>
+        /// Gets the current query statistics for server-side processing time. The statistics are stateful.
+        /// </summary>
+        ///
+        /// <returns>The current TimingInformation statistics.</returns>
+        public TimingInformation GetTimingInformation()
+        {
+            return this.timingInformation;
         }
     }
 }